GCP/PubSub/CloudFunctions/pubsub-receiver/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore>=2.1.3
